requirements.txt
supabase==2.3.4
bcrypt==4.1.2
PyJWT==2.8.0
pandas==2.2.0
openpyxl==3.1.2
Pillow==10.2.0
